requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic-settings==2.8.1
requests-html==0.10.0
beautifulsoup4==4.13.3
PyJWT==2.10.1
jinja2==3.1.6
pytest==8.3.5
httpx==0.28.1
lxml_html_clean==0.4.1